requirements.txt
blinker==1.8.2
click==8.1.3
Flask==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.0
python-dotenv==1.0.0
